+++ src/components/user-reservation/confirmation/AccountDetails.jsx
@@ -22,7 +22,6 @@ const AccountDetails = () => {
         <div className="card-body">
           <h2 className="card-title">
             <svg
-              xmlns="http://www.w3.org/2000/svg"
               fill="none"
               viewBox="0 0 24 24"
               strokeWidth={1.5}
