+++ src/components/user-reservation/reservation/TimeSlot.jsx
@@ -1,22 +1,27 @@
 import PropTypes from "prop-types";
 import { useDispatch } from "react-redux";
-import { setTime, setPage } from "../../../utils/slice/reservationSlice";
+import { setTime } from "../../../utils/slice/reservationSlice";
 import { convertDBTime } from "../../../lib/timeSlotUtil";
+import { useNavigate } from 'react-router-dom';
 
 const TimeSlot = ({ info }) => {
+  const navigate = useNavigate();
   const dispatch = useDispatch();
+
   return (
     <div
       className={`${
         info.available ? "cursor-pointer" : "pointer-events-none grayscale"
-      } p-1 m-1 w-full border-2 border-transparent text-center`}
+      } p-4 mx-auto w-[90%] border-2 flex flex-row items-center border-transparent border-b-beige-s hover:invert`}
       onClick={() => {
         dispatch(setTime(convertDBTime(info.time)));
-        dispatch(setPage("confirmation"));
+        navigate('/reservation/confirmation');
       }}
     >
-      {info.time}
-      <div className="text-[.5em] text-slate-400 mt-1 mx-auto w-fit">
+      <div>
+        {info.time}
+      </div>
+      <div className="text-[.5em] text-slate-400 ml-auto w-fit flex items-center">
         Remaining Spots:
         <span className="text-xs ml-1 text-slate-600">{info.spots}</span>
       </div>
