+++ src/containers/UserReservation.jsx
@@ -1,16 +1,10 @@
-import { useSelector } from "react-redux";
-import Search from "../components/user-reservation/search/Search.jsx";
-import Reservation from "../components/user-reservation/reservation/Reservation.jsx";
-import Confirmation from "../components/user-reservation/confirmation/Confirmation.jsx";
+import { Outlet } from 'react-router-dom';
 
 const UserReservation = () => {
-  const { page } = useSelector((state) => state.reservation);
 
   return (
     <div className="flex flex-col justify-center items-center">
-      {page === "search" && <Search />}
-      {page === "reservation" && <Reservation />}
-      {page === "confirmation" && <Confirmation />}
+      <Outlet />
     </div>
   );
 };
