+++ src/utils/getReservationsByDate.js
@@ -4,7 +4,7 @@ const getReservationsByDate = async (garage_id, date) => {
   /* Given garage_id and date, get count of reservations per hour */
   try {
     const { data } = await axios.get(
-      `http://localhost:3001/reservations/${garage_id}?date=${date}`
+      `http://localhost:3001/reservations/garage/${garage_id}?date=${date}`
     );
     return data;
   } catch (error) {
